+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.24.8] - 2024-11-14
+
 ### Changed
 
 - Reconnections now use instantaneous filament velocity information by default.